Git diff advanced args for better diff
backwards compatible for now

+++ bash_scripts/lib_generate_fluxo_diff_files.sh
@@ -1,7 +1,49 @@
 #!/usr/bin/env bash 
 set -e
 
+FLUXO_DIFF_HELP_MESSAGE="\
+GIT-FLUXODIFF
+
+$(tput bold)USAGE$(tput sgr0)
+
+  git fluxo diff 
+  git fluxo diff [ <some-fluxo-branch> | -a | --all ]
+  git fluxo diff [ <any-branch>..<any-other-branch> ]
+  git fluxo diff ( -o | --output-files ) [<output-directory>]
+
+  git fluxo diff [ ( -o | --output-files ) <output-directory> ] [ <some-fluxo-branch> | -a | --all ] [ <any-branch>..<any-other-branch> ]
+
+  git fluxo diff -h | --help
+
+$(tput bold)PARAMS$(tput sgr0)
+
+  <some-fluxo-branch>
+      is a fluxo branch you want to get the diff off. Defaults to current branch.
+      The '_fluxo_bnranches' file will be read to determine wich branch is the previous fluxo branch.
+
+  <any-branch>..<any-other-branch>
+      is any set of branches you want to diff using _fluxo_* file rules
+
+$(tput bold)ACTIONS$(tput sgr0)
+
+  -h | --help      show detailed instructions.
+  -a | --all       diff all branches following '_fluxo_branches' file.
+  -o | --output    generates diff files in <output-directory>.
+"
+
+function print_diff_full_usage {
+  printf '\n%s\n' "$FLUXO_DIFF_HELP_MESSAGE"
+}
+
+function print_diff_usage_and_die {
+  printf '\n%s\n' "$FLUXO_DIFF_HELP_MESSAGE"
+  exit 1
+}
+
 . $(cd "$(dirname "$0")" && pwd)"/lib_style.sh"
+. $(cd "$(dirname "$0")" && pwd)"/lib_parse_fluxo_branches_file.sh"
+. $(cd "$(dirname "$0")" && pwd)"/lib_show_fluxo.sh"
+. $(cd "$(dirname "$0")" && pwd)"/lib_view.sh"
 
 function read_fluxo_file {
     local FILE_NAME="$1"
@@ -18,23 +60,125 @@ function read_fluxo_file {
     fi
 }
 
+function find_previous_for_from {
+    local _item_name="$1"
+    local _list="$2"
+
+    local _current_item_name _previous_item_name
+    while read -r _current_item_name; do
+        if [ "$_current_item_name" = "$_item_name" ]; then
+            printf %s "$_previous_item_name"
+            return
+        fi
+        _previous_item_name="$_current_item_name"
+    done <<< "$(printf '%b' "$_list")"
+}
+
+function parse_args {
+    local branches_arg branches_var_name="$1"
+    local output_directory_arg output_directory_var_name="$2"
+
+    shift
+    shift
+
+    local should_diff_all_branches=-1
+    local first_branch_arg second_branch_arg
+
+    local BRANCHES_ARGS=""
+
+    while [ $# -gt 0 ]; do
+        case "$1" in
+            -h|--help)
+                print_full_usage | less -XRF
+                clear
+                exit $?
+                ;;
+            -o|--output-files)
+                shift
+                if [ -z "$1" ]; then
+                    printf '\n%s\n' "No output folder dir provided"
+                    print_diff_usage_and_die
+                fi
+                output_directory_arg="$1"
+                shift
+                ;;
+            -a|--all)
+                should_diff_all_branches=0
+                shift
+                ;;
+            *)
+                BRANCHES_ARGS+="$1 "
+                shift
+                ;;
+        esac
+    done
+
+    [ ! -z "$BRANCHES_ARGS" ] && set -- "${BRANCHES_ARGS%% }"
+
+    if [ $# -gt 0 ] && [ $should_diff_all_branches -eq 0 ]; then
+        printf '\n%s\n%s\n' "You selected both '--all' branches option and passed a specific branch argument '$@' to diff" "You must choose either all branches or a specific set of branches to diff"
+        print_diff_usage_and_die
+    fi
+
+    if [ $# -gt 2 ]; then
+        printf '\n%s\n%s\n' "More than one branch argument provided" "To diff between to branches you must use the syntax: branch1..branch2"
+        print_diff_usage_and_die
+    fi
+
+    local fluxo_branches="$(show_fluxo --existent --raw)"
+
+    if [ $should_diff_all_branches -eq 0 ]; then
+        branches_arg="$fluxo_branches"
+    else
+        local first_branch_arg="$(printf '%s' "$1" | awk '{ split($0,list,/\.\./); print list[1] }')"
+        local second_branch_arg="$(printf '%s' "$1" | awk '{ split($0,list,/\.\./); print list[2] }')"
+
+        if [ -z "$first_branch_arg" ]; then
+            local current_branch="$(git rev-parse --abbrev-ref HEAD)"
+            
+            local previous_branch
+            previous_branch="$(find_previous_for_from "$current_branch" "$fluxo_branches")"
+
+            if [ -z "$previous_branch" ]; then
+                printf '\n%s\n\n' "$(view_errorline) Current branch '$current_branch' is not a fluxo branch"
+                exit 1
+            fi
+
+            branches_arg="$previous_branch\\n$current_branch"
+        elif [ ! -z "$first_branch_arg" ] && [ -z "$second_branch_arg" ]; then
+            local previous_branch
+            previous_branch="$(find_previous_for_from "$first_branch_arg" "$fluxo_branches")"
+
+            if [ -z "$previous_branch" ]; then
+                printf '\n%s\n\n' "$(view_errorline) Argument'$first_branch_arg' is not a fluxo branch"
+                exit 1
+            fi
+
+            branches_arg="$previous_branch\\n$first_branch_arg"
+        elif [ ! -z "$first_branch_arg" ] && [ ! -z "$second_branch_arg" ]; then
+            branches_arg="$first_branch_arg\\n$second_branch_arg"
+        else
+            print_diff_usage_and_die
+        fi
+    fi
+
+    eval "$branches_var_name=\$branches_arg"
+    eval "$output_directory_var_name=\$output_directory_arg"
+}
+
 function generate_fluxo_diff_files {
-    [ -z "$1" ] && local dest_folder="_fluxo_diff_files" || local dest_folder=$1
-    local project_dir="$(pwd)"
+    local branches 
+    local output_directory
+
+    parse_args branches output_directory "$@"
+
+    [ -z "$output_directory" ] && local dest_folder="_fluxo_diff_files" || local dest_folder="$output_directory"
     
     local tmp_folder=$(mktemp -d)
     function remove_tmp_folder {
         rm -rf "$1"
     }
     trap "remove_tmp_folder $tmp_folder" EXIT
-
-    local branches="$(show_fluxo --existent --raw)"
-
-    status="$?"
-    if [ $status != 0 ]; then
-        echo -e "$branches\n"
-        exit $status
-    fi
 
     local root_dir="$(read_fluxo_file "_fluxo_root")"
     local ignored_files="$(read_fluxo_file "_fluxo_ignore")"
@@ -91,9 +235,11 @@ function generate_fluxo_diff_files {
         fi
     done
 
-    rm -r "$dest_folder" 2> /dev/null
-    mkdir -p $dest_folder
-    mv $tmp_folder/* $dest_folder
+    mkdir -p "$dest_folder"_new
+    mv $tmp_folder/* "$dest_folder"_new
+    [ -r "$dest_folder" ] && rm -r "$dest_folder" 2> /dev/null
+
+    mv "$dest_folder"_new "$dest_folder"
 
     echo
     echo -e "Created $(style $BOLD$PURPLE $qt_files) diff files in $(style $UNDERLINE$CYAN `pwd $dest_folder`/$dest_folder/)"
